QA fixes Signed-off-by: Joey Smith <jsmith@webinertia.net>
