--- src/Container/DriverInterfaceFactoryFactoryInterface.php
@@ -1,9 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace PhpDb\Container;
-
-interface DriverInterfaceFactoryFactoryInterface extends FactoryFactoryInterface
-{
-}
