--- src/Container/PlatformInterfaceFactoryFactoryInterface.php
@@ -1,9 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace PhpDb\Container;
-
-interface PlatformInterfaceFactoryFactoryInterface extends FactoryFactoryInterface
-{
-}
