+++ test/unit/ConfigProviderTest.php
@@ -6,16 +6,17 @@ namespace PhpDbTest;
 
 use PhpDb\Adapter;
 use PhpDb\ConfigProvider;
-use PhpDb\Container\AdapterAbstractServiceFactory;
+use PhpDb\Container\AbstractAdapterInterfaceFactory;
 use PHPUnit\Framework\TestCase;
 
 class ConfigProviderTest extends TestCase
 {
     /** @phpstan-var array{'dependencies': array{abstract_factories: list<class-string>, aliases: array<class-string, class-string>}} */
     private array $config = [
-        'dependencies' => [
+        Adapter\AdapterInterface::class => [],
+        'dependencies'                  => [
             'abstract_factories' => [
-                AdapterAbstractServiceFactory::class,
+                AbstractAdapterInterfaceFactory::class,
             ],
             'aliases'            => [
                 Adapter\AdapterInterface::class => Adapter\Adapter::class,
